pinger/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
prometheus-client==0.21.0
httpx==0.27.2
pyyaml==6.0.2